Migrate Bootstrap column classes: replace BS 2 classes with BS 3 classes
As a first step in the Bootstrap migration, change all 'span*' to 'col-md-*'.
There will be much fine-tuning of columns for optimal responsiveness later.
(That is, 'col-md-*' might not always be the best class to use; in some cases
it might be 'col-sm-*' or 'col-xs-*', but that will be easier to determine
on a page-by-page basis after the pages are halfway presentable.)